% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_package}
\alias{add_package}
\title{Add a package to a repository}
\usage{
add_package(repo = ".", package, branch = "master", type = "source",
  token = NULL)
}
\arguments{
\item{repo}{character - A repository URL or local directory with a remote.}

\item{package}{character - An R package in a directory, a git remote to a
repo containing an R package, or a link to a package in a CRAN repo.}

\item{branch}{character - The branch of the gitCRAN repository to write to.}

\item{type}{character - What directory to group the repository under. If not
"source", the default, then it compiles the package as a binary but still
adds it to the {type}/src/contrib/ directory.}

\item{token}{character - The personal access token (PAT) for reading/writing
to the repository.}
}
\description{
To use this, the user must have a personal access token (PAT) from the git
provider must be generated for reading and writing.
}
\examples{
\dontrun{
  add_package(
    "github.com/thomascjohnson/testCRAN",
    "github.com/thomascjohnson/gitCRAN",
    branch = "dev",
    type = "cflinuxfs3"
  )
}
}
