% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_packages.R
\name{add_packages}
\alias{add_packages}
\title{Add packages to a repository}
\usage{
add_packages(
  repo = ".",
  packages,
  branch = "master",
  subrepo = "source",
  build = FALSE,
  image = NULL,
  token = NULL,
  save_subrepo_config = TRUE,
  update_image = TRUE,
  lib_dir = NULL
)
}
\arguments{
\item{repo}{character - A repository URL or local directory with a remote.}

\item{packages}{character - R packages in a directory, a git remote to a
repo containing an R package, or a link to a package in a CRAN repo.}

\item{branch}{character - The branch of the gitCRAN repository to write to.}

\item{subrepo}{character - What directory to group the repository under. If
not "source", the default, then it compiles the package as a binary but
still adds it to the {type}/src/contrib/ directory.}

\item{image}{character - A Docker image name, either}

\item{token}{character - The personal access token (PAT) for reading/writing
to the repository.}

\item{save_subrepo_config}{logical - Whether or not to save the build and
image options to a .gitCRAN-config file in the subrepo root.}
}
\description{
To use this, the user must have a personal access token (PAT) from the git
provider must be generated for reading and writing.
}
\examples{
\dontrun{
  add_packages(
    "github.com/thomascjohnson/testCRAN",
    "github.com/thomascjohnson/gitCRAN",
    branch = "dev",
    name = "cflinuxfs3"
  )
}
}
