% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{build_package}
\alias{build_package}
\title{Returns location of built package}
\usage{
build_package(package, build, container, package_dir, available_packages)
}
\description{
Returns location of built package
}
