% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-request.R
\name{package_request_pipeline}
\alias{package_request_pipeline}
\title{Handle Package Requests from Issues Automatically}
\usage{
package_request_pipeline(
  owner = Sys.getenv("GITCRAN_REPO_OWNER"),
  gh_repository = Sys.getenv("GITCRAN_REPO"),
  labels = Sys.getenv("GITCRAN_LABELS", "package-request"),
  state = Sys.getenv("GITCRAN_STATE", "open"),
  username = Sys.getenv("GITHUB_USER"),
  token = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{owner}{character - Github organization/user name, defaults to the
environment variable GITCRAN_REPO_OWNER}

\item{gh_repository}{character, the repository name, defaults to the
environment variable GITCRAN_REPO}

\item{labels}{character vector - Github issue labels to search for package
requests, defaults to the environment variable GITCRAN_LABELS}

\item{state}{character - Issue state, "open" or "closed", defaults to the
environment variable GITCRAN_STATE or "open"}

\item{username}{character - Github username that has access to repository
issues, defaults to the environment variable GITHUB_USER}

\item{token}{character - Github Personal Access TOKEN (PAT) for the provided
username with repository read and write permissions, defaults to the
environment variable GITHUB_PAT}
}
\description{
Given the
}
