% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_token}
\alias{set_token}
\title{Set Provider Personal Access Tokens}
\usage{
set_token(provider, token)
}
\arguments{
\item{provider}{character - Which git provider the token is being set for.
Available providers: github, bitbucket, gitlab}

\item{token}{character - the token value}
}
\description{
The token must be a token that can read and write to repositories. This will
create a file named .gitCRAN in the home directory. The user can also set
the env var GIT_CRAN_CONFIG to the desired config path and it will instead
write there.
}
\details{
Personal Access Tokens
See the following links to create personal access tokens for the different providers:
\itemize{
  \item \href{https://help.github.com/en/github/authenticating-to-github/creating-a-personal-access-token-for-the-command-line}{Github}
  \item \href{https://confluence.atlassian.com/bitbucketserver/personal-access-tokens-939515499.html}{Bitbucket}
  \item \href{https://docs.gitlab.com/ce/user/profile/personal_access_tokens.html}{Gitlab}
}
}
\examples{
\dontrun{
  set_token("Github", "not_my_actual_pat")
}
}
